Rewrite to YARP based reverse proxy
