--- it.bz.noi.community-api.Tests/.env.example
@@ -1,5 +1,0 @@
-CLIENT_ID=
-CLIENT_SECRET=
-SERVICE_SCOPE=
-SERVICE_URL=
-TENANT_ID=
