--- it.bz.noi.community-api.Tests/FunctionTest.cs
@@ -1,20 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-using Xunit;
-
-using it.bz.noi.community_api;
-using dotenv.net;
-
-namespace it.bz.noi.community_api.Tests
-{
-    public class FunctionTest
-    {
-        public FunctionTest()
-        {
-            DotEnv.Load();
-        }
-    }
-}
