+++ it.bz.noi.community-api.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30114.105
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "it.bz.noi.community-api", "it.bz.noi.community-api\it.bz.noi.community-api.csproj", "{1137E2B7-B364-4174-B0A7-5FBB30EB7D36}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "it.bz.noi.community-api.Tests", "it.bz.noi.community-api.Tests\it.bz.noi.community-api.Tests.csproj", "{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}"
 EndProject
 Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{48BEDD14-D366-4553-AD99-C35DDAF28FF2}"
 EndProject
@@ -31,18 +29,6 @@ Global
 		{1137E2B7-B364-4174-B0A7-5FBB30EB7D36}.Release|x64.Build.0 = Release|Any CPU
 		{1137E2B7-B364-4174-B0A7-5FBB30EB7D36}.Release|x86.ActiveCfg = Release|Any CPU
 		{1137E2B7-B364-4174-B0A7-5FBB30EB7D36}.Release|x86.Build.0 = Release|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Debug|x64.Build.0 = Debug|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Debug|x86.Build.0 = Debug|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Release|x64.ActiveCfg = Release|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Release|x64.Build.0 = Release|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Release|x86.ActiveCfg = Release|Any CPU
-		{781DFBD4-4889-4446-B2C0-1717B9D1EDFC}.Release|x86.Build.0 = Release|Any CPU
 		{48BEDD14-D366-4553-AD99-C35DDAF28FF2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{48BEDD14-D366-4553-AD99-C35DDAF28FF2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{48BEDD14-D366-4553-AD99-C35DDAF28FF2}.Debug|x64.ActiveCfg = Debug|Any CPU
