+++ it.bz.noi.community-api/appsettings.json
@@ -1,0 +1,23 @@
+{
+ "AllowedHosts": "*",
+ "ReverseProxy": {
+   "Routes": {
+     "route1" : {
+       "ClusterId": "cluster1",
+       "AuthorizationPolicy": "noi-auth",
+       "Match": {
+         "Path": "{**catch-all}"
+       }
+     }
+   },
+   "Clusters": {
+     "cluster1": {
+       "Destinations": {
+         "destination1": {
+           "Address": "https://noitest.crm4.dynamics.com/api/data/v9.2/"
+         }
+       }
+     }
+   }
+ }
+}
